Refactor logic with mixin.
